Revert "Pass the branch name to the kernel tests."
This isn't actually useful, because on many branches, $BRANCH is
not the actual name of the branch, in order to avoid merge
conflicts. This approach will also lead, in the long term, to
cluttering the test files with skip statements.

Instead of taking this approach, we will branch kernel/tests to
ensure that each kernel branch runs the kernel tests that are
appropriate to it.

This reverts commit 710746c697821b6ec08a1de6cf450666f7855547.