build: add buildinfo.py for compilation info
Meant to be invoked as an EXTRA_CMD to build.config.  Expects $OUT_DIR
env var to be set.  Produces a json file named compile.json in $OUT_DIR.
This file is the complete list of -I, -D, and a few other flags that
we're used during the compilation of this kernel.  This gives us the
subset of source files, and info we need to produce a proper AST of the
code in our kernel for static analysis.

usage: OUT_DIR=/path/to/out/branch/ python buildinfo.py

Bug: 31904467
Change-Id: Ia0fbbfb2d9346150fde8e42bdd1a3feb69df0b13