Combines two changes for checkpatch.sh.
This is a convenience for subsequent cherrypicks.

Change-Id: If633c907db9209f81544fd93259b2b15ad118c9a

Adds --repo_prop. Checks range of changes after given sha1.

--repo_prop is useful when you know the name of the repo before hand,
but not the sha1 (i.e. as in presubmit).

It's better to check the range of patches after the given sha1, since
the value given by repo.prop may refer to the oldest change in a stack.

As as a side effect, the patches are written to files. This enables
checkpatch.pl to also warn about errors in the change description,
whereas it (apparently) fails to do so if the patch is passed on stdin.

Change-Id: I7ac0b82140608b80409bb685575c48ea882e5c1d