Narrows range of patches, report full results.
Change call to git format-patch to only return patch for given sha1,
instead of sha1..HEAD. This is what the user intends in most cases and
has the added benefit of simplifying the code, since there is guaranteed
to be a single patch.

Report a summary of the errors to build.log, while including the
complete output in the dist dir.

Shows error types by default. Ignore GERRIT_CHANGE_ID lint errors.

BUG:33207015
Change-Id: I62a53981935cd46e68a4ddac418713de38496d05