Build modules in separate out dir, strip modules
This change fixes three issues:
 - Build out-of-tree modules in separate output directory (O=)
 - Strip kernel modules

When building out-of-tree modules, we want to store object files (.o)
and other output files in a directory that is different from the source
directory. For in-tree kernel builds, this can be achieved by specifying
O= on the make command line. For out-of-tree kernel modules, it is
somewhat more difficult to achieve: make needs to be invoked in the
kernel source directory, and the variable M needs to be set to a
*relative* path and point to the source of the out-of-tree module. The
variable O needs to point to the output directory of the kernel build.
The output files of the out-of-tree module will then end up in
$(O)/$(M).

It turned out that "make modules_install" digitally signs the kernel
modules before installing them. The explicit signing of kernel modules
can hence be removed from build.sh.

The previous version of this change had two issues:

Some existing build.config files depend on ${OUT_DIR} to point to the
kernel output directory. The previous version of this change had renamed
OUT_DIR to KERNEL_OUT_DIR and therefore broke existing scripts.

scripts/sign-file has been called unconditionally for all .ko files even
if there was no sign-file executable.

Change-Id: I09da1bc48b4e8b7f55e10b9bdd7d207fcee17c0a
Signed-off-by: Daniel Mentz <danielmentz@google.com>