Build modules in separate out dir, strip modules
This change fixes three issues:
 - Build out-of-tree modules in separate output directory (O=)
 - Strip kernel modules

When building out-of-tree modules, we want to store object files (.o)
and other output files in a directory that is different from the source
directory. For in-tree kernel builds, this can be achieved by specifying
O= on the make command line. For out-of-tree kernel modules, it is
somewhat more difficult to achieve: make needs to be invoked in the
kernel source directory, and the variable M needs to be set to a
*relative* path and point to the source of the out-of-tree module. The
variable O needs to point to the output directory of the kernel build.
The output files of the out-of-tree module will then end up in
$(O)/$(M).

Change-Id: If2f720959729d08e2d04ce40c7f291e443231489
Signed-off-by: Daniel Mentz <danielmentz@google.com>