Use clang for out-of-tree modules if requested
A non-default C compiler can be specified in build.config by setting the
variable CC.  This is usually done if the kernel should be compiled with
clang (i.e.  CC=clang). When building out-of-tree modules, this setting
has previously been ignored. As a result, out-of-tree modules were built
with GCC instead of clang.

Change-Id: Icaed54db3882237a065be3ceea6311a4feca6626
Signed-off-by: Daniel Mentz <danielmentz@google.com>